+++ catboost/libs/fstr/shap_values.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "shap_prepared_trees.h"
+
 #include <catboost/libs/data/data_provider.h>
 #include <catboost/libs/model/model.h>
 #include <catboost/private/libs/options/enums.h>
@@ -11,22 +13,6 @@
 #include <util/system/types.h>
 #include <util/ysaveload.h>
 
-
-struct TShapValue {
-    int Feature = -1;
-    TVector<double> Value;
-
-public:
-    TShapValue() = default;
-
-    TShapValue(int feature, int approxDimension)
-        : Feature(feature)
-        , Value(approxDimension)
-    {
-    }
-
-    Y_SAVELOAD_DEFINE(Feature, Value);
-};
 
 struct TFixedFeatureParams {
     enum class EMode {
@@ -60,42 +46,6 @@ public:
     );
 };
 
-struct TShapPreparedTrees {
-    TVector<TVector<TVector<TShapValue>>> ShapValuesByLeafForAllTrees; // [treeIdx][leafIdx][shapFeature] trees * 2^d * d
-    TVector<TVector<double>> MeanValuesForAllTrees;
-    TVector<double> AverageApproxByTree;
-    TVector<int> BinFeatureCombinationClass;
-    TVector<TVector<int>> CombinationClassFeatures;
-    bool CalcShapValuesByLeafForAllTrees;
-    bool CalcInternalValues;
-    TVector<double> LeafWeightsForAllTrees;
-    TVector<TVector<TVector<double>>> SubtreeWeightsForAllTrees;
-
-public:
-    TShapPreparedTrees() = default;
-
-    TShapPreparedTrees(
-        const TVector<TVector<TVector<TShapValue>>>& shapValuesByLeafForAllTrees,
-        const TVector<TVector<double>>& meanValuesForAllTrees
-    )
-        : ShapValuesByLeafForAllTrees(shapValuesByLeafForAllTrees)
-        , MeanValuesForAllTrees(meanValuesForAllTrees)
-    {
-    }
-
-    Y_SAVELOAD_DEFINE(
-        ShapValuesByLeafForAllTrees,
-        MeanValuesForAllTrees,
-        AverageApproxByTree,
-        BinFeatureCombinationClass,
-        CombinationClassFeatures,
-        CalcShapValuesByLeafForAllTrees,
-        CalcInternalValues,
-        LeafWeightsForAllTrees,
-        SubtreeWeightsForAllTrees
-    );
-};
-
 void CalcShapValuesForDocumentMulti(
     const TFullModel& model,
     const TShapPreparedTrees& preparedTrees,
@@ -115,15 +65,6 @@ void CalcShapValuesForDocumentMulti(
     TConstArrayRef<NCB::NModelEvaluation::TCalcerIndexType> docIndices,
     size_t documentIdx,
     TVector<TVector<double>>* shapValues
-);
-
-TShapPreparedTrees PrepareTrees(const TFullModel& model, NPar::TLocalExecutor* localExecutor);
-TShapPreparedTrees PrepareTrees(
-    const TFullModel& model,
-    const NCB::TDataProvider* dataset, // can be nullptr if model has LeafWeights
-    EPreCalcShapValues mode,
-    NPar::TLocalExecutor* localExecutor,
-    bool calcInternalValues = false
 );
 
 void CalcShapValuesByLeaf(
