+++ catboost/libs/fstr/ya.make
@@ -7,8 +7,9 @@ SRCS(
     compare_documents.cpp
     feature_str.cpp
     output_fstr.cpp
-    shap_values.cpp
     shap_interaction_values.cpp
+    shap_values.cpp
+    shap_prepared_trees.cpp
     util.cpp
 )
